Hiding room HTML content and appending the Jitsi iFrame in order to avoid to close and open again the room to recover messages.
Not using 'oob_url' in order to avoid having a message and link which are not corresponding to the action of opening the meeting in a new tab.

Fixing the link which permits to open meeting into the room which was not always working from the room of the user that initialized the meeting.