Modifying the jitsimeet iframe behavior since the version 8.0.0 of ConverseJS. The IFrame is now inserted into the DOM so that it is not removed on component view rerendering. This new behavior allows the user to switch on other chat room while maintaining the Jitsi Meet active in background.
A new plugin parameter is also added (for display modes other than the overlayed one) : 'jitsimeet_head_display_toggle'. If enabled, by setting true, the header of the chatroom from which the jitsi meet has been started keep displayed above the iframe. If the user double click on it, then the display of jitsi meet iframe is toggled between hidden and displayed state whithout stopping the meet. By this way, while the meet is in progress, the user can hide the meet iframe to write something into the chat and display again the meet.
In overlayed mode, this parameter is ignored and the display toggling is actived.